requirements.txt
tqdm==4.65.0
scikit-image==0.19.3
opencv-python==4.6.0.66
ConfigArgParse==1.7
lpips==0.1.4
imageio-ffmpeg==0.4.8
kornia==0.6.12
tensorboard==2.13.0
jaxtyping==0.2.19
rich==13.4.2
transformers==4.31.0
accelerate==0.21.0
diffusers==0.19.0
xformers==0.0.22
einops==0.6.1
pytorch-lightning==2.1.0
plyfile==1.0.1
mmcv==2.1.0
mmcv-full==1.7.1
yapf==0.40.1




